solver/requirements.txt
Pillow
PyAutoGUI
PyGetWindow
